+++ src/App.js
@@ -18,12 +18,7 @@ function App() {
   const[date,setDate] = useState();
   const staticInfo = "";
 
-  // for toggling info content
-  const [showContent, setShowContent] = useState(false);
-  const toggleMode = e => {
-    e.preventDefault();
-    setShowContent(!showContent);
-  };
+
 
   useEffect(() => {
     axios
@@ -52,9 +47,7 @@ function App() {
   return (
     <div className="App">
       <Header date = {date} />
-      <ImgCard imgURL ={photo} imgTitle = {title} explanation = {info} showContent = {showContent}/>
-      <Button type="primary" onClick={toggleMode}>Show more info</Button>
-
+      <ImgCard imgURL ={photo} imgTitle = {title} explanation = {info} />
     </div>
 
   );
